Create titlecase util function
Create titlecase util function
